whiterabbit: implement retry on timeout
As a side effect this commit simplifies a little the WR state machine
by factorizing all timeout events in the white rabbit handshake.

On slave timeouts, we now go to PPS_SLAVE and non-wr mode (not PPS_LISTENING,
which was wrong).

Finally, this adds a missing calib_pattern_disable(), in the timeout
path of wr_resp_calib_req().

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>